requirements.txt
aioredis>=2.0.1
annotated-types>=0.7.0
anyio>=4.11.0
async-timeout>=5.0.1
asyncpg>=0.31.0
bcrypt==4.0.1
certifi>=2025.11.12
cffi>=2.0.0
click>=8.3.1
cryptography>=46.0.3
Deprecated>=1.3.1
distro>=1.9.0
dnspython>=2.8.0
ecdsa>=0.19.1
email-validator>=2.3.0
fastapi>=0.122.0
greenlet>=3.2.4
h11>=0.16.0
httpcore>=1.0.9
httptools>=0.7.1
httpx>=0.28.1
idna>=3.11
jiter>=0.12.0
limits>=5.6.0
openai>=2.8.1
openpyxl>=3.1.5
packaging>=25.0
pandas>=2.3.3
passlib>=1.7.4
pyasn1>=0.6.1
pycparser>=2.23
pydantic>=2.12.5
pydantic-settings>=2.12.0
pydantic_core>=2.41.5
python-dotenv>=1.2.1
python-jose>=3.5.0
python-multipart>=0.0.20
PyYAML>=6.0.3
redis>=7.1.0
rsa>=4.9.1
six>=1.17.0
slowapi>=0.1.9
sniffio>=1.3.1
SQLAlchemy>=2.0.44
starlette>=0.50.0
tqdm>=4.67.1
typing-inspection>=0.4.2
typing_extensions>=4.15.0
uvicorn>=0.38.0
uvloop>=0.22.1
watchfiles>=1.1.1
prometheus_client>=0.23.1
websockets>=15.0.1
wrapt>=2.0.1
xlsxwriter>=3.2.0